Restructure folders and files
