+++ frontend/src/routes/AppWrapper.tsx
@@ -1,0 +1,10 @@
+import { Outlet } from "react-router-dom";
+import NavBar from "../navigation/NavBar";
+
+// wraps all app pages
+export default function AppWrapper() {
+  return (<>
+    <NavBar />
+    <Outlet />
+  </>)
+}
