+++ frontend/src/routes/PublicOnlyRoutes.tsx
@@ -1,7 +1,7 @@
 import { Navigate, useLocation, Outlet } from 'react-router-dom';
 import { useAuth } from '../utils/hooks';
 
-export default function PublicOnly() {
+export default function PublicOnlyRoutes() {
   const auth = useAuth();
   const location = useLocation();
   if (auth.authState) {
